Separated main UI components into their own files.
