+++ app/src/main/java/com/rohankhayech/choona/view/screens/ConfigureTuningScreen.kt
@@ -59,6 +59,8 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.rohankhayech.choona.R
 import com.rohankhayech.choona.model.tuning.Tunings
+import com.rohankhayech.choona.view.components.StringControls
+import com.rohankhayech.choona.view.components.TuningSelector
 import com.rohankhayech.choona.view.theme.AppTheme
 import com.rohankhayech.music.Tuning
